\name{L3-package}
\alias{L3-package}
\alias{L3}
\docType{package}
\title{
\packageTitle{L3}
}
\description{
\packageDescription{L3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{L3}
\packageIndices{L3}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{L3}

Maintainer: \packageMaintainer{L3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
